- use 4 byte entries for the script.
We no longer store pointers in here.

+++ source/games/duke/src/gamedef.cpp
@@ -61,7 +61,7 @@ int checking_ifelse;
 extern char tempbuf[];
 extern int* labelcode;
 
-TArray<intptr_t> ScriptCode;
+TArray<int> ScriptCode;
 
 //---------------------------------------------------------------------------
 //
